crud
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@5096 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormSigElementDefEdit.cs
@@ -503,7 +503,7 @@ namespace OpenDental{
 			}
 			ElementCur.SigElementType=(SignalElementType)listType.SelectedIndex;
 			ElementCur.SigText=textSigText.Text;
-			ElementCur.LightRow=PIn.Int(textLightRow.Text);
+			ElementCur.LightRow=PIn.Byte(textLightRow.Text);
 			ElementCur.LightColor=butColor.BackColor;
 			if(IsNew){
 				SigElementDefs.Insert(ElementCur);
